test: 최단 거리를 구하는 인수 테스트를 작성하라

+++ src/main/java/nextstep/subway/applicaion/dto/PathResponse.java
@@ -5,21 +5,21 @@ import java.util.List;
 public class PathResponse {
 
     private List<StationResponse> stations;
-    private Double distance;
+    private int distance;
 
     public PathResponse() {
     }
 
     public PathResponse(List<StationResponse> stations, Double distance) {
         this.stations = stations;
-        this.distance = distance;
+        this.distance = distance.intValue();
     }
 
     public List<StationResponse> getStations() {
         return stations;
     }
 
-    public Double getDistance() {
+    public int getDistance() {
         return distance;
     }
 }
